requirements.txt
# Pin versions used by this project; adjust if your environment differs.
Django>=6.0,<7.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
supabase>=2.0,<3.0
python-dotenv>=1.0,<2.0
